requirements.txt
appdirs==1.4.3
asgiref==3.2.3
attrs==19.3.0
beautifulsoup4==4.9.0
black==19.10b0
blis==0.4.1
cachetools==4.0.0
catalogue==1.0.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
cryptography==2.8
cycler==0.10.0
cymem==2.0.3
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.2
django-chartjs==2.0.0
django-crispy-forms==1.8.1
django-extensions==2.2.8
django-heroku==0.3.1
django-tables2==2.2.1
feedparser==5.2.1
google-api-core==1.16.0
google-cloud-language==1.3.0
googleapis-common-protos==1.51.0
grpcio==1.26.0
gunicorn==20.0.4
idna==2.8
importlib-metadata==1.5.0
javascript==2015.32.59
kiwisolver==1.1.0
matplotlib==3.1.3
murmurhash==1.0.2
numpy==1.18.1
oauthlib==3.1.0
pathspec==0.7.0
Pillow==7.0.0
plac==1.1.3
preshed==3.0.2
protobuf==3.11.2
psycopg2==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.6
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.3
regex==2020.1.8
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.14.0
soupsieve==2.0
spacy==2.2.3
sqlparse==0.3.0
srsly==1.0.1
thinc==7.3.1
toml==0.10.0
tqdm==4.42.1
typed-ast==1.4.1
Unidecode==1.1.1
urllib3==1.25.8
wasabi==0.6.0
whitenoise==5.0.1
zipp==2.2.0
